test(http): split rate-limit tests into dedicated test file (#1062)
Move the `describe "failed auth rate limiting"` block and the
`rate limiting disabled` test from http_router_test.exs into a new
file http_router_rate_limit_test.exs, keeping http_router_test.exs
under 800 lines (776 → 776, from 902). All 794 mcp_server tests pass.
